move validators in modules
